requirements.txt
fastapi
<<<<<<< HEAD
uvicorn[standard]
sqlalchemy
alembic
python-dotenv
psycopg2-binary
pyjwt
pydantic
=======
uvicorn
sqlalchemy
alembic
psycopg2-binary
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
pytest
httpx
python-dotenv
requests
email-validator
>>>>>>> c6a823c (Projeto semi finalizado)
